requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.7
groq==0.4.2
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
Werkzeug==2.2.3
gunicorn==21.2.0
